Move event_handler function to another file again
