+++ src/event_handler.rs
@@ -5,48 +5,14 @@ use crate::config::key_action::KeyAction;
 use crate::config::key_press::{KeyPress, Modifier};
 use crate::Config;
 use evdev::uinput::VirtualDevice;
-use evdev::{Device, EventType, InputEvent, Key};
+use evdev::{EventType, InputEvent, Key};
 use lazy_static::lazy_static;
 use log::debug;
-use nix::sys::select::select;
-use nix::sys::select::FdSet;
 use std::collections::HashMap;
 use std::error::Error;
-use std::os::unix::io::AsRawFd;
 use std::time::Instant;
 
-pub fn event_loop(
-    mut input_devices: Vec<Device>,
-    output_device: VirtualDevice,
-    config: &Config,
-) -> Result<(), Box<dyn Error>> {
-    let mut handler = EventHandler::new(output_device);
-    loop {
-        let readable_fds = select_readable(&input_devices)?;
-        for input_device in &mut input_devices {
-            if readable_fds.contains(input_device.as_raw_fd()) {
-                for event in input_device.fetch_events()? {
-                    if event.event_type() == EventType::KEY {
-                        handler.on_event(event, config)?;
-                    } else {
-                        handler.send_event(event)?;
-                    }
-                }
-            }
-        }
-    }
-}
-
-fn select_readable(devices: &Vec<Device>) -> Result<FdSet, Box<dyn Error>> {
-    let mut read_fds = FdSet::new();
-    for device in devices {
-        read_fds.insert(device.as_raw_fd());
-    }
-    select(None, &mut read_fds, None, None, None)?;
-    return Ok(read_fds);
-}
-
-struct EventHandler {
+pub struct EventHandler {
     device: VirtualDevice,
     wm_client: WMClient,
     multi_purpose_keys: HashMap<Key, MultiPurposeKeyState>,
@@ -59,7 +25,7 @@ struct EventHandler {
 }
 
 impl EventHandler {
-    fn new(device: VirtualDevice) -> EventHandler {
+    pub fn new(device: VirtualDevice) -> EventHandler {
         EventHandler {
             device,
             wm_client: build_client(),
@@ -74,7 +40,7 @@ impl EventHandler {
     }
 
     // Handle EventType::KEY
-    fn on_event(&mut self, event: InputEvent, config: &Config) -> Result<(), Box<dyn Error>> {
+    pub fn on_event(&mut self, event: InputEvent, config: &Config) -> Result<(), Box<dyn Error>> {
         self.application_cache = None; // expire cache
         let key = Key::new(event.code());
         debug!("=> {}: {:?}", event.value(), &key);
@@ -104,7 +70,7 @@ impl EventHandler {
         Ok(())
     }
 
-    fn send_event(&mut self, event: InputEvent) -> std::io::Result<()> {
+    pub fn send_event(&mut self, event: InputEvent) -> std::io::Result<()> {
         if event.event_type() == EventType::KEY {
             debug!("{}: {:?}", event.value(), Key::new(event.code()))
         }
