+++ src/main.rs
@@ -1,8 +1,14 @@
 use crate::config::Config;
 use crate::device::{input_devices, output_device};
-use crate::event_handler::event_loop;
+use crate::event_handler::EventHandler;
+use evdev::uinput::VirtualDevice;
+use evdev::{Device, EventType};
 use getopts::Options;
+use nix::sys::select::select;
+use nix::sys::select::FdSet;
 use std::env;
+use std::error::Error;
+use std::os::unix::io::AsRawFd;
 use std::process::exit;
 
 extern crate getopts;
@@ -11,16 +17,6 @@ mod client;
 mod config;
 mod device;
 mod event_handler;
-
-fn usage(program: &str, opts: Options) -> String {
-    let brief = format!("Usage: {} CONFIG [options]", program);
-    opts.usage(&brief)
-}
-
-fn abort(message: &str) -> ! {
-    println!("{}", message);
-    exit(1);
-}
 
 fn main() {
     env_logger::init();
@@ -62,4 +58,45 @@ fn main() {
     if let Err(e) = event_loop(input_devices, output_device, &config) {
         abort(&format!("Error: {}", e));
     }
+}
+
+fn event_loop(
+    mut input_devices: Vec<Device>,
+    output_device: VirtualDevice,
+    config: &Config,
+) -> Result<(), Box<dyn Error>> {
+    let mut handler = EventHandler::new(output_device);
+    loop {
+        let readable_fds = select_readable(&input_devices)?;
+        for input_device in &mut input_devices {
+            if readable_fds.contains(input_device.as_raw_fd()) {
+                for event in input_device.fetch_events()? {
+                    if event.event_type() == EventType::KEY {
+                        handler.on_event(event, config)?;
+                    } else {
+                        handler.send_event(event)?;
+                    }
+                }
+            }
+        }
+    }
+}
+
+fn select_readable(devices: &Vec<Device>) -> Result<FdSet, Box<dyn Error>> {
+    let mut read_fds = FdSet::new();
+    for device in devices {
+        read_fds.insert(device.as_raw_fd());
+    }
+    select(None, &mut read_fds, None, None, None)?;
+    return Ok(read_fds);
+}
+
+fn usage(program: &str, opts: Options) -> String {
+    let brief = format!("Usage: {} CONFIG [options]", program);
+    opts.usage(&brief)
+}
+
+fn abort(message: &str) -> ! {
+    println!("{}", message);
+    exit(1);
 }
